Automatically backed up by Learn

+++ lib/cash_register.rb
@@ -33,7 +33,7 @@ class CashRegister
   def apply_discount
     if @discount!=0
        
-    @discount = @discount/100.to_i
+    @discount = @discount/100.to_f
     @total = @total-(@total*@discount)
     
        # total
